# gestimowebfront-end/src/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger API pour GESTIMOWEB REST API",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger GESTIMOWEB REST API",
    "contact" : {
      "name" : "bzadata",
      "email" : "michel.bossoh@outlook.fr"
    }
  },
  "basePath" : "http://localhost:8282/",
  "paths" : {
...
    "gestimoweb/api/v1/suiviedepense/allSuivieDepense/{idAgence}" : {
      "get" : {
        "operationId" : "getAllEncaissementSuivieDepenseParAgence",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SuivieDepenseDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/countSuiviNonCloturerParCaisseEtChapitreAvantDate/{datePriseEnCompteEncaii}/{idCaiss}/{idChapitre}" : {
      "get" : {
        "operationId" : "countSuiviNonCloturerParCaisseEtChapitreAvantDate",
        "parameters" : [ {
          "name" : "datePriseEnCompteEncaii",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "idCaiss",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idChapitre",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/getSuivieDepenseByCodeTransaction/{codeTransaction}" : {
      "get" : {
        "operationId" : "getSuivieDepenseByCodeTransaction",
        "parameters" : [ {
          "name" : "codeTransaction",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SuivieDepenseDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/getSuivieDepenseById/{id}" : {
      "get" : {
        "operationId" : "getSuivieDepenseById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SuivieDepenseDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/listSortieDeuxDate/{idAgence}/{debut}/{fin}" : {
      "get" : {
        "operationId" : "listSortieDeuxDate",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "debut",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "fin",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SuivieDepenseDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/listSuiviDepenseNonCloturerParCaisseEtChapitrAvantDate/{idcaisse}/{dateDepriseEnCompte}/{idChapitre}" : {
      "get" : {
        "operationId" : "listSuiviDepenseNonCloturerParCaisseEtChapitrAvantDate",
        "parameters" : [ {
          "name" : "idcaisse",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "dateDepriseEnCompte",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "idChapitre",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SuivieDepenseDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/listSuiviDepenseNonCloturerParCaisseEtChapitreEntreDeuxDate/{idcaisse}/{dateDebut}/{dateFin}/{idChapitre}" : {
      "get" : {
        "operationId" : "listSuiviDepenseNonCloturerParCaisseEtChapitreEntreDeuxDate",
        "parameters" : [ {
          "name" : "idcaisse",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "dateDebut",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "dateFin",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "idChapitre",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SuivieDepenseDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/saveSuivieDepense" : {
      "post" : {
        "operationId" : "saveSuivieDepense",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SuivieDepenseDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SuivieDepenseDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/suprimerSuiviParId/{id}/{idAgence}" : {
      "post" : {
        "operationId" : "suprimerSuiviParId",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SuivieDepenseDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/suiviedepense/totalSortieDeuxDate/{idAgence}/{debut}/{fin}" : {
      "get" : {
        "operationId" : "totalSortieDeuxDate",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "debut",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "fin",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SuivieDepenseEncaisPeriodeDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/villa/all/{idAgence}" : {
      "get" : {
        "operationId" : "findAllVilla",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VillaDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/villa/alllibre/{idAgence}" : {
      "get" : {
        "operationId" : "findAllVillaLibre",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VillaDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/villa/findVillaById/{id}" : {
      "get" : {
        "operationId" : "findVillaById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VillaDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/villa/save" : {
      "post" : {
        "operationId" : "saveVilla",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/VillaDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VillaDto"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "SuivieDepenseDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateEncaissement" : {
          "type" : "string",
          "format" : "date"
        },
        "designation" : {
          "type" : "string"
        },
        "codeTransaction" : {
          "type" : "string"
        },
        "montantDepense" : {
          "type" : "number",
          "format" : "double"
        },
        "modePaiement" : {
          "type" : "string",
          "enum" : [ "ESPESE_MAGISER", "ESPECE_SEVEINVEST", "CHEQUE_ECOBANK_MAGISER", "CHEQUE_GTBANK_SEVEINVEST", "VIREMENT_ECOBANK_MAGISER", "VIREMENT_GTBANK_SEVEINVEST", "MOBILE_MONEY_MAGISER", "MOBILE_MONEY_SEVEINVEST" ]
        },
        "operationType" : {
          "type" : "string",
          "enum" : [ "CREDIT", "DEBIT" ]
        },
        "cloturerSuivi" : {
          "type" : "string"
        },
        "idChapitre" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "SuivieDepenseEncaisPeriodeDto" : {
      "type" : "object",
      "properties" : {
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateDebutEncaissement" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFinEncaissement" : {
          "type" : "string",
          "format" : "date"
        },
        "designation" : {
          "type" : "string"
        },
        "codeTransaction" : {
          "type" : "string"
        },
        "totalMontantDepense" : {
          "type" : "number",
          "format" : "double"
        }
      }
    },
...
    "VillaDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nbrePieceVilla" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbrChambreVilla" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbrSalonVilla" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbrSalleEauVilla" : {
          "type" : "integer",
          "format" : "int32"
        },
        "numVilla" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeAbrvBienImmobilier" : {
          "type" : "string"
        },
        "nomCompletBienImmobilier" : {
          "type" : "string"
        },
        "nomBaptiserBienImmobilier" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "superficieBien" : {
          "type" : "number",
          "format" : "double"
        },
        "bienMeublerResidence" : {
          "type" : "boolean"
        },
        "idSite" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idUtilisateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "proprietaire" : {
          "type" : "string"
        },
        "idChapitre" : {
          "type" : "integer",
          "format" : "int64"
        },
        "occupied" : {
          "type" : "boolean"
        }
      }
    },
...
  }
}